Cleanup and renaming.
