+++ src/teehr/__init__.py
@@ -1,3 +1,3 @@
 from teehr.evaluation.evaluation import Evaluation  # noqa
-from teehr.models.metrics.metrics_models import Metrics  # noqa
+from teehr.models.metrics.metric_models import Metrics  # noqa
 from teehr.models.metrics.metric_enums import Operators  # noqa
